requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-embed-video==1.1.2
django-extensions==2.1.2
django-pagedown==1.0.6
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.2.0
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
selenium==3.14.1
six==1.11.0
stripe==2.4.0
urllib3==1.23
wrapt==1.10.11
